Revert "rofi-register-login"
